Fix VUID-08740 in pipeline feedback creation tests
Writing to gl_PointSize in tessellation and geometry shaders
is only allowed if the device feature is supported. Not all
implementations do, so we remove the write to that output in
tessellation and geometry shaders. This way no coverage is lost
and the VUID is avoided.

Affects:
dEQP-VK.pipeline.monolithic.creation_feedback.graphics_tests.*

VK-GL-CTS: 4337
Components: Vulkan
Change-Id: Ic49539ed33b8f8a3eb06c2d0ade3a6064a5a1e21