Pass ModifierProp instead of raw modifier value in some modifier tests
We are going to need other properties of a modifier later. They are
stored in ModifierProp outside of the main body of export-import case.

It seems the Vulkan spec 1.3 doesn't define a method to query properties
of a single modifer (VkDrmFormatModifierPropertiesEXT). They can only be
obtained when querying 'all' modifiers in a list
(VkDrmFormatModifierPropertiesListEXT).

Passing down ModifierProp avoids another round of query and search in the
export-import test case.

Signed-off-by: Jianxun Zhang <jianxun.zhang@intel.com>