Faster mustpass files generation
Rewritten mustpass txt files generation to be 3-4x times faster (i.e. from 10 minutes to 3 minutes).
The new implementation is also simpler than the previous one. It reads the input files and, line by line, it matches each test case with the include and exclude filters.
The result is simpler code and faster execution.
Filters are still divided into two main groups:
 - simple test case names, which need to match litterally
 - patterns with wildcards

For simple test case names, we build a tree structure so that we can match the test case names by splitting it by the dot character.

Because the previous implementation was storing all the input test cases in a dictionary, the order of the test cases in the output files was not deterministic.
The new algorithm preserves the order of the input files, albeit split into multiple output files.

The new txt mustpass files, included in this commit, are identical to the previous ones, if we compare them after sorting both of them alphabetically.

VK-GL-CTS issue: 4791

This change does not affect any test.

Change-Id: I6df504ad354f27e2facb60657c50a47d32e939dc