Add setVisible for VulkanWindowDirectDrm
I forgot to add this function when I introduced VulkanWindowDirectDrm because it didn't make sense for the platform, but now some tests are using it.
It should just be a NOOP. As far as I can tell, only these tests currently use the setVisible() function.

Component: Vulkan
VK-GL-CTS issue: 4806

Affected tests:
    dEQP-VK.wsi.direct_drm.full_screen_exclusive.*
    dEQP-VK.wsi.direct_drm.maintenance1.*

Change-Id: Icbb8039eb9dcc58f7d28b1f2db97fdfbb3cc44da