Fix dEQP-VK.shader_object.misc.state.pipeline.*.color_blend.enabled
This fixes the test when running on implementation that does not support
`extendedDynamicState3ColorBlendEquation`.

It looks that when the feature is not supported, the test uses different
blending state:
 - static blending state has ZERO dst factor
 - dynamic blending state (used only if extendedDynamicState3ColorBlendEquation
   is supported) uses ONE fst factor.

This change adapts the color verification to expect 'whiteColor'=vec4(0.75)
if ZERO dst factor is used.

Components: Vulkan

VK-GL-CTS Issue: 4868

Affects: dEQP-VK.shader_object.misc.state.pipeline.*.color_blend.enabled
Change-Id: Ia9747b0fba2916bdf559c13682f966a39cab749a