Fix missing memory barrier with GL4 shader image
The affected GL4 test writes to an image with imageStore, and then in a
second draw samples from that image as a texture. According to the
OpenGL specification section 7.13.2 ("Shader Memory Access
Synchronization"), this requires a memory barrier to flush caches in
between.

The corresponding GLES test has the memory barrier (although the type of
barrier is wrong...), it is only the GL test that lacks it. I am not
sure how that asymmetry happened, the bug dates back to the initial
versions of these tests added back in 2014.

Add the required memory barrier, fixing the test particularly for
implementations with incoherent image caches.

Affects:

KHR-GL46.shader_image_load_store.advanced-sync-imageAccess2

Components: OpenGL

VK-GL-CTS issue: 4867

Change-Id: Idd0404cf3b93043bbf9776594800684b9c1030ce