Test frag shader invocations with non-flat colors
Implementations are allowed to reuse the same output for multiple
fragments, but they can't do so if the color is not known to be constant
at runtime or if the frag shader has side effects.

This commit adds non-flat color shading variants and verify the number
of frag shader invocations matches the pixel count.

New and affected tests:
dEQP-VK.query_pool.frag_invocations.*

Components: Vulkan
VK-GL-CTS issue: 4851

Change-Id: I9712a8cf82f5c1b65f267c6a0460c27c47ac48da