Do not rely on undefined behaviour in subgroup tests
The identity for max/min operations is +/-inf. When testing these
operations, don't attempt to check the result when no elements of the
subgroup have been included in the operation, as this produces a
comparison against +/-inf. The SPIR-V compiler or the GPU ALU is allowed
to assume infinities don't occur in the program, and can optimise the
comparison to false, which currently will make the test fail.

Affects:
dEQP-VK.subgroups.*.subgroup*max_*
dEQP-VK.subgroups.*.subgroup*min_*

Component: Vulkan
VK-GL-CTS issue: 4812

Change-Id: Iae93fa9949d59a781a44c2a36e32e9b1a963ae6e