Fix device free in extension_duplicates test
The api.extension_duplicates.device.* tests use the createUncheckedDevice function
to create a device used in testing. Due to the implementation of the VulkanSC variant
of the CTS, this will result in a device being created once the first time the
test is run in the main process, and a second device being created when the test is
run again in a subrocess. However, the call used to clean up the device will only
call into an ICD in the subprocess run of the test, leaving an unfree'd device from
the main process.

This change updates the destroy device call used in the tests. This aligns with other
tests using the createUncheckedDevice function, such that the created device is freed
in both the main process and subprocess test runs.

Affects tests: dEQP-VKSC.api.extension_duplicates.device.*

Components: VulkanSC
VK-GL-CTS issue: 4970

Change-Id: Ieae888644838a6898b71539d72e8c187ac101c46