Add coverage for mutable descriptor type list being out of range.
As per the specification, if VkMutableDescriptorTypeCreateInfoEXT::
pMutableDescriptorTypeLists[i] is out of range, the descriptor pool
allocates enough memory to be able to allocate a
VK_DESCRIPTOR_TYPE_MUTABLE_EXT descriptor with any supported
VkDescriptorType as a mutable descriptor. To cover this case, tests are
added that create bindings of zero or more non-mutable descriptors
followed by one or more mutable descriptors. The number of descriptor
pool size entries are equal to the number of all bindings in the test but
the number of type lists in VkMutableDescriptorTypeCreateInfoEXT is
equal to the number of the non-mutable descriptors only. This forces the
pool to determine the allocation of memory for mutable descriptors by
itself.

New tests:
dEQP-VK.binding_model.mutable_descriptor.misc.mutable_type_out_of_range_*

Affects:
dEQP-VK.binding_model.mutable_descriptor.*

Components: Vulkan
VK-GL-CTS issue: 4764

Change-Id: Iee75724bb25498013cdc1677b6df093b4caa4714