Fix: Test PVRTC extension and always use power-of-two dimensions
Public PR: https://github.com/KhronosGroup/VK-GL-CTS/pull/451

The CTS tests always use random image sizes for testing images, which was not valid for PVRTC1 textures. With 1.3.279 a VUID was added to require power-of-two dimensions for PVRTC1 textures, as discussed in KhronosGroup/Vulkan-Docs#1991.

Also, this fixes a bug introduced with 4c7e4e9, as it required the extension to be present for the PVRTC formats to be tested, but it was never added as a extension to be enabled, so PVRTC textures were never tested.

Components: Vulkan

Affects:
dEQP-VK.api.invariance.random

Change-Id: I89884bed683b796ccc9c793ba419455f842e71ba