Fix float_controls2 tests assuming RTE rounding
The FMA tests check that 0.1 * 10.0 - 1.0 == 0.0 which fails on RTZ as
a result of a rounding error.
Change the tests to check 0.1 * 23.0 - 2.3 == 0.0 to allow them to
work on both RTE and RTZ.

Some tests add or multiply two large inputs and expect the result to
overflow to infinity.
This works as expected for RTE but will produce the max finite value
for RTZ.
Change the tests to require the shaderRoundingModeRTEFloat{16,32,64}
property and use RoundingModeRTE in the shader.

Affects: dEQP-VK.spirv_assembly.instruction.*.float_controls2.*.input_args.*
Components: Vulkan
VK-GL-CTS issue: 4962

Change-Id: I34492b26aff70409d78d51976353072c4a8a8c08