Add test log parser for new android activity
This change adds four new signatures in ContainerFormatParser
To process the logs generated by the new Android
activity introduced in https://gerrit.khronos.org/c/vk-gl-cts/+/14242:

1) beginTestRunParamsCollection
2) endTestRunParamsCollection
3) beginTestRunParams
4) endTestRunParams

This change also adds a new class type KhronosCTSTestLogListener,
a subclass of TestLogListener. The new KhronosCTSTestLogListener
receives the processed log from the ContainerFormatParser,
and notifies the android instrumentation process
(code for the instrumentation process is in future changes) based on
the information received in the processed log. That way the instrumentation
process will know if the new android activity has finished the four tasks.
The instrumentation process is also able to get the test run parameters,
which is needed for the instrumentation process to launch the
Khronos GLES 3.2 CTS in Android test infra.

This change also modifies the function signature of
TestLogParser::parse() to take a TestLogListener pointer, instead of
a reference. So that both parent class TestLogListner instance,
and the child class KhronosCTSTestLogListener instance, can be
passed to the same function.

Affected tests: None

Components: Framework

VK-GL-CTS issue: 5024

Change-Id: I2ffad0684ae8be2220d1ad283f12b0d535babb0e