Fix wsi maintenance1 scaling tests
The wsi maintenance1 scaling tests were breaking the Vulkan
specification regarding VkSurfacePresentModeEXT. Namely, the
presentMode specified in VkSurfacePresentModeEXT must be a value
reported by vkGetPhysicalDeviceSurfacePresentModesKHR for the
specified surface.

Here the tests can potentially query for the capabilities of a present
mode that the swapchain implementation does not support.

This is fixed by skipping the test if the present mode is not
supported.

Affects:
dEQP-VK.wsi.*.maintenance1.scaling.*

Components: Vulkan

VK-GL-CTS issue: 5161

Change-Id: Ic22165a5480508fc65581d0b8cdb5151c6ada360