Fix return code check for vkGetPhysicalDeviceVideoFormatPropertiesKHR
There are a couple of places in the video test utility code where the
return code from vkGetPhysicalDeviceVideoFormatPropertiesKHR is examined
to signal an "unsupported" case to the calling code: if the return code
is VK_ERROR_FORMAT_NOT_SUPPORTED, the implementation does not support
the input combination of image format and image usage for the video
profile provided in the API call.

However, this check is incorrect because VK_ERROR_FORMAT_NOT_SUPPORTED
is not a valid return code for this API call. There are other
video-specific return codes for errors, which this API call can return.
While the code can be written to check for one for those return codecs,
it is easier to check that the API call has not returned VK_SUCCESS.
(VK_INCOMPLETE is another non-error return code from this API call, but
is not expected to be returned in current usage).

Components: Vulkan

VK-GL-CTS issue: 5142

Affected tests:
dEQP-VK.video.encode.*

Change-Id: I60ad166ea53ef5f004023b31480c6ad3caea1622