Add glEnable/glDisable for clip distances
The test ClipDistancesValuePassing checks to see if the clip distances
are sent properly from vertex shader to fragment shader. This test
fails, however, because there is no glEnable/glDisable for the clip
distances that are used. The glEnable/glDisable are used and checked
in other tests in that file, but are not present in this method.

Affects:
KHR-GLES3.clip_distance.coverage

Components: OpenGL ES

Google bug: 341780527
VK-GL-CTS issue: 5160

Change-Id: I47643f34937c722e1a320445ad786da639b17842