Fix incorrect usages of DE_NULL
DE_NULL was supposed to be a pointer, but having been defined as 0, it
was misused in many places.

This change defines DE_NULL as nullptr and fixes ensuing compile errors.
DE_NULL was mostly used in place of VK_NULL_HANDLE, which was cleaned up
in previous changes.

Components: Vulkan
VK-GL-CTS issue: 4652

Change-Id: I31a5964ec73018a3bcd9d8b8cc3432e280bc1dac