correct output for direct_state_access.buffers_clear when failure occurs
For R8* formats, the data is interpreted as non-numeric, so it messes
with the log output. This change ensures the numbers are produced.

Current outcome for R8* formats failure:
The test of ClearNamedBufferData has failed for internalformat . Cleared buffer data is equal to [

After the fix:
The test of ClearNamedBufferData has failed for internalformat . Cleared buffer data is equal to [0], but [5] was expected.

Affects:
KHR-GL46.direct_state_access.buffers_clear

Change-Id: I9e9d771ba129464824df258d5a7066ac797b4c3a
Signed-off-by: David Heidelberg <david@ixit.cz>