Fix multiple memory leaks in glcts
Test cases have a deinit() method that is called once the test case has
been entered (i.e. the test has already gone through the iterate()
method). However, de deinit() method is not called when the test is not
entered. In particular, this means that any resources allocated in the
test case constructor and only freed in deinit() may leak because the
test case may be constructed as part of creating the test hierarchy,
allocating resources, but if the test selection does not include that
particular test, the case will not be entered and deinit() will not be
called to free those resources.

There are several ways to work around this issue:

1. Call deinit() for that particular test class in the destructor of the
test class. When doing this, we must be careful to reset pointers and
resource values to indicate they're no longer allocated after freeing
them, because if the test is entered, deinit() will be called twice (one
explicitly when leaving the test case and once more when destroying the
test hierarchy.

2. Move resource allocation to iterate(), so resources never leak due to
test case creation. This is also more efficient as we do not waste time
allocating resources that will never be used.

3. Only use members with automatic destructors. This will free the
resources as part of the class destructor in any case. It can be used in
combination with (2), but is harder to apply when resources are
allocated via the GL API.

For this fix, option 2 has been selected in most cases and option 1 in
some others.

In addition, point 3 has been applied (by using an std::unique_ptr) to a
few unrelated leaks that were forgetting to free memory of some objects.

Also in addition, this commit also fixes a glBufferData call that was
using the wrong array size.

Affects:
KHR*.geometry_shader.*
KHR*.shader_group_vote.*

VK-GL-CTS issue: 5214
VK-GL-CTS issue: 5219
Components: OpenGL

Change-Id: I0f732b5db8e7c5524f80d37f9684cb68ba2ce5cd