Fix offsets in pixel sampling edge tests
This commit reduces slightly the offset used to test edge sampling.
The original value of 0.499 was indistinguishable from 0.5 in the
case that the texture width is 256 and the implementation was only
accurate to about 1 part in 10^5 (the minimum required by the spec).

Affects:
dEQP-VK.pipeline.monolithic.sampler.exact_sampling.*.edge_right

Components: Vulkan

VK-GL-CTS issue: 5168

Change-Id: Icd2356009be146aa2f093efbf287ed2127f7c7fa