Fix a VkSampler use-after-free in the VK_EXT_descriptor_buffer capture/replay tests
The tests capture all of the descriptors in the first iteration and only
re-capture in the second iteration for things where we're explicitly
testing capture/replay.  For the sampled image tests, there is a sampler
that is not part of the capture/replay cycle and we're creating it,
capturing the descriptor, and then destroying and re-creating it before
we draw.  The draw is done using the descriptor from the first sampler.
We could re-capture the descriptor or, more easily, we could just not
re-create the sampler and keep the one from the first iteration around.

Components: Vulkan

VK-GL-CTS issue: 5247

Affected tests:
dEQP-VK.binding_model.descriptor_buffer.capture_replay.*sampled_image*

Change-Id: I4136856349c5f1df8e8d07039241823541751c9d