Resolve workaround for instance entry points
Generation of vkInstanceDriverImpl.inl was done with an
assumption that instance entry points that have aliases
are Vulkan 1.1 entry points promoted from extensions.
This was added back in the Vulkan 1.1 times and for later
vulkan versions it was patched to produce correct results.
This change removes workaround and applies proper search
when the entry point was promoted.

Components: Framework

VK-GL-CTS issue: 4695

Affects:
dEQP-VK.*

Change-Id: Ieaf8669e6ec30e75c45d2433bbe8e7f2ff35a334