build: allow manual configuration of WAYLAND_PROTOCOLS_DIR
pkg_get_variable blindly overwrites the target variable, even if the
user has already provided a value for it on the command line. Without
this change, it is difficult to use a version of wayland-protocols other
than the system-provided one. This is especially important when
cross-compiling or developing new protocols.

No functional change

Change-Id: I4c1f26a448833fc8ae66b43e9eb264249c32ef55