Make autogen for VK, VK SC consistent
The autogen was excluding Vulkan core feature structs from the list of
extension structs to be filled in but was not excluding the SC versions.
This meant that the SC versions had to be explicitly skipped in the
code. It makes much more sense for all the core feature structs to be
handled in the same way.

Components: VulkanSC
Affects: dEQP-VK-SC.*

Change-Id: I80d09630f0cc5003c33e1b5d8d0b297238591dff