Fix frag_coord_conventions test when not supported
The test init was trying to store whether the test was supported so that
it could return NotSupported at the appropriate time. However, the test
setup, before the deferred exit, still required the implementation to
support the features. Correctly skip the setup if the test is not
supported so that the original return will mark the test as unsupported.

Components: OpenGL
VK-GL-CTS issue: 4005
Affects: KHR-GLES31.core.frag_coord_conventions.multisample

Change-Id: I99b6388a9a38e66b1c27f795994ea873cbedfa49