Avoid comparing garbage memory in descriptor buffer capture-replay tests
The descriptor info for a storage image was being obtained to a 256-byte
buffer that may exceed the space needed for a storage image according to
the properties. This size was being passed as dataSize in a call to
vkGetDescriptorEXT, hitting VUID-vkGetDescriptorEXT-dataSize-08125.

In addition, this was being done twice with two different buffers and
then the full buffer contents were being compared for equality. However,
an implementation may only fill the first storageImageDescriptorSize
bytes of the buffer, making the comparison fail due to garbage bytes at
the end of the buffer.

Affects:
dEQP-VK.binding_model.descriptor_buffer.capture_replay.*

Components: Vulkan
VK-GL-CTS issue: 5408

Change-Id: I5df1bc27fec4fb2e4c6e92554e5e5c97f0ca2236