Fix synchronization in multiview tests
For the dynamic rendering depth/stencil case, the barrier for the
depth/stencil image was missing.

For dynamic rendering without depth/stencil, the dst access needed
COLOR_ATTACHMENT_READ because the contents of the color attachment are
preserved between subpasses.

For renderpass, VK_ACCESS_INPUT_ATTACHMENT_READ was incorrect because
even in the test that is use input attachments, the input attachment is
a separate image from the color attachment. Dependencies for reusing the
depth/stencil and color attachments were missing.

The src stage for renderpass depth/stencil was also changed, because
depth writes happen in LATE_FRAGMENT_TESTS, not EARLY_FRAGMENT_TESTS.

Components: Vulkan

Affects:
dEQP-VK.multiview.*

VK-GL-CTS issue: 5479

Change-Id: If4114fb77cf970b64ceec4684bef6a40cf27a857