Fix loading absolute path libs with LD_LIBRARY_PATH
Commit 5aa5b0809f055f36fc384644925c384596b2f7cd changed the way
libraries are dynamically loaded when `LD_LIBRARY_PATH` is set to only
take the filename into account and not the complete path. This seems to
be overlooked code meant for testing as the commit is about video
testing and not library loading.

Revert the file to its previous state to fix loading libraries that are
specified with absolute paths.

Fixes GitHub issue #479.