Coverage for different operations and helper invocations
Each test draws the same triangle twice. Result of the first draw is
copied to buffer and passed as input to the second draw. With the
first draw we just want to know which pixels are under the triangle.
In the second draw fwidth is called on the value read from the input
buffer (result of first draw).
Test expects to get one of four allowed values - cler color around
the triangle, 0 inside triangle and one of two values on triangle
edges that depend on clear and render values used in the first draw.
By verifying all four values we will know if values that were
grabbed (from input buffer) in helper invocations were read
correctly.

Components: Vulkan

VK-GL-CTS issue: 5371

Affects:
dEQP-VK.glsl.helper_invocations.*

Change-Id: Ia008231492e22c6420db9b0832f4bdd6a1b0ff9d