Fix wsi maintenance1 present mode query tests
When a system has multiple Vulkan ICDs installed, such as AMDVLK and Mesa RADV, the Vulkan loader is responsible for querying all instance extensions supported by these drivers. However, when invoking vkGetPhysicalDeviceSurfaceCapabilities2KHR, the loader sometimes only dispatches the function to AMDVLK. Unfortunately, this driver does not support the instance extension VK_EXT_surface_maintenance1, which prevents it from processing the new structure VkSurfacePresentModeCompatibilityEXT that is extended within VkSurfaceCapabilities2KHR.

According to the Vulkan specification VkSurfacePresentModeEXT, if pPresentModes is NULL, then the number of present modes compatible with the one specified in VkSurfacePresentModeEXT is returned in presentModeCount. Given that a present mode passed to VkSurfacePresentModeEXT must be compatible with itself, receiving a value less than 1 indicates that the extension is not implemented on the ICD being used.

Based on this, the situation where compatibility.presentModeCount is less than 1 should be considered as a case of unsupport rather than a query failure.


Affects:
dEQP-VK.wsi.xlib.maintenance1.present_modes.*.query

Components: Vulkan
VK-GL-CTS issue: 5457

Change-Id: Ibaa0f2b0a35fe7bdf98c1346cc736f66c84337db