Test pairing a frag shader with PrimitiveID with mesh and vert shaders
There is a group of mesh shader tests that try using a fragment shader
that's shared between a mesh shader and a vertex shader. This "sharing"
is achieved using GPL, with a common fragment shader lib and a mesh or
vert pre-rasterization shader lib. To make the tests more interesting,
both the mesh and vertex shader set gl_Layer, which is defined in the
mesh spec as a perprimitiveEXT built-in while in the vertex shader it's
a regular built-in, and use gl_Layer in the fragment shader as well.

This commit adds the following coverage for this group of tests:

* Using PrimitiveID, which is interesting because it's not only
  perprimitiveEXT in the mesh shader, but also you don't need to set it
  from the vertex shader, creating a different problematic case.

* Unlinked SPIR-V shader object variants of these tests.

* Attempting both command sequences, using mesh shading first or last.

Affects and new tests:
dEQP-VK.mesh_shader.ext.smoke.*shared*

Components: Vulkan
VK-GL-CTS issue: 5460

Change-Id: I0bc1190c72fbacf2a2542b0f4dcfba1c7b2ed81d