Simplify feature struct consistency checking
Rather than doing the work of the test as best we can if the features
are not supported then not checking the result at the end, skip the test
if the feature struct being tested is not allowed. This also makes it
easier to debug whether the tests are really testing anything or not.

Components: Vulkan

Change-Id: Ibed747e034aca0fb75bf1ec532bb9ad09a6ae092