Test alpha to coverage with alpha to one
To make sure these operations run in the right order, the tests cover
the framebuffer in geometry but make the output color have alpha 0.0.

If alpha to coverage is run correctly before alpha to one, this should
result in no coverage and the framebuffer ends up with the clear color.
If alpha to coverage is run after alpha to one, this results in coverage
and the framebuffer ends up with the geometry color using alpha 1.0.

New tests:
dEQP-VK.pipeline.*.multisample.a2c_with_a2one.*

Components: Vulkan
VK-GL-CTS issue: 5489

Change-Id: Ic13637ca4d1ba0b56806e7ac9fbf7828bf4bd2c0