Lower dimensions for image concurrent copy tests
The dimensions for the image in these tests is up to 256,256,32.
We split them into regions with random dimensions between (1,32). If we
take the average of 16, that results in 512 regions. In the "host" and
"multiple" configuration, we then create a thread for each region to
concurrently copy.

On Linux, a common stack size is 8mb.  That means we want to use 4.2gb,
i.e. the limit for 32-bit machines.  When I looked at a test, I saw it
try to call pthread_create 672 times, but failed on 430.

So, we should lower the image size.

Components: Vulkan

VK-GL-CTS issue: 5510

Affects:
dEQP-VK.image.concurrent_copy.*

Change-Id: I9dfa375da6c437ef49f4b7f4894007bf2c5a72d6