Add tests creating pipelines/shaders with no queues.
Part of VK_KHR_maintenance9. The tests create pipelines or shaders on a device
with no queues, query them with a pipeline cache, pipeline binary, or shader
binary, and recreate them on a device with queues and render with them to
verify they work as intended.

New tests: dEQP-VK.pipeline.no_queues.*

Components: Vulkan
Change-Id: I7c5ebd08494a6fd5d69da686917379716681e0cf