Fix precision bug in fragment area for FDM tests
The FDM fragment area is calculated as 1.0 / densityValue, where the
density value comes from the FDM attachment (red for the width of the
area, green for the height). The driver is allowed to use any area it
supports as long as the total area is not above the area that would be
calculated the way mentioned above.

In these tests, one of the parameters is the target fragment area, which
can be something like 2x2. To calculate the density value that needs to
be stored in the attachment to achieve the target area, the tests divide
1.0 by the target area size. For 2x2, this results in 1.0 / 2.0 = 0.5
for both the width and height values (red and green). However, because
the attachment format is R8G8_UNORM, the driver cannot store 0.5 exactly
in the attachment. It may store 8-bit values 127 or 128. If it rounds up
to 128, the resulting area is

    1.0 / (128 / 255) = 1.0 / 0.5019607843137255 = 1.9921875.

And, as per the specification, the implementation cannot use 2x2 because
the area would be bigger than 1.99x1.99.

To fix the issue, we apply a small delta to the density value depending
on the attachment format to guarantee that the target area is always a
valid choice for the driver.

Affects:
dEQP-VK.*.fragment_density_map.*

Components: Vulkan
VK-GL-CTS issue: 4135

Change-Id: Ifb9045cf15c1c4f90f2bf47e76363157e8c5721e