Refactor Vulkan video encode tests
This is part of an effort to streamline legacy code in video
encode/decode tests, with the goal of simplifying their
maintenance and evolution.

This change refactors `VideoEncodeTestInstance::iterate()`
method, which itself is more than 1000 LoC. It breaks down
logical blocks into new methods, organizing a bit the code
in the process.

Affects:
dEQP-VK.video.encode.*

Components: Vulkan

Change-Id: I22a830e4f967e4a53f7746a8ae0b965aabd4661c