Linux compile failure when commenting out some test additions
This fixes the problem of dependencies of linked libraries.
When all packages were built together the problem did not
occur because the linker saw all the required dependencies,
but when the linker was separate the linker complained about
not finding them and the build failed. To fix this issue some
packages introduce tiny targets which are included as
dependencies in the packages that required them.
As a result, we can attach packages separately or mix them
freely.

This patch also fixes the problem of the vk::createShader(...)
function not being linked into the vk(sc)utilnoshader libraries.
This function was marked as inline in the translation unit so
only the header declaration was visble and available.
This function was also rewritten verbatim in the shader_object
package which resulted in multiple definitions of the same func.
Eventually the definition was removed from the shader_objec
package and packages that use it, including pipeline and compute
which use the one defined in vk(sc)utilnoshader.

Components: Vulkan, VulkanSC

VK-GL-CTS Issue: 5550

Affects: None, only build process

Change-Id: I8f65dbe52e68bf3ce509a92d4e942bac7be7a4ef