fetch_sources.py: Fix missing usage of fallback URL
There's support for trying both http and ssh git
URLs if both are specified, but the fallback URL
was being ignored when attempting to checkout the
specified tag/commit in a repo, only when cloning
a fresh repo.

This change fixes that by retrying the checkout
step in the exception handler and re-raising the
exception if no fallback URL is available.

Change-Id: I49c2030f988737c42edb1d3f79f7c5c2267c577a