Add tests with additional commands on secondary command buffers
Some drivers implement secondary command buffers with a "capture/replay"
scheme, where the commands are recorded to a buffer on the host and then
replayed at ExecuteCommands time. This requires care when handling
Vulkan commands that have pointers in them, as 1. manual code is
sometimes required, and 2. the lifetime of the object pointed to might
be shorter than the secondary command buffer for some structures,
leading to a use-after-free.

New tests:
dEQP-VK.api.command_buffers.secondary_push*

Components: Vulkan
VK-GL-CTS issue: 5175

Change-Id: I2e23456cb828c975b9f26a19576be4f5a926f593