Fix assertion in 64-bit blit tests
Apparently no vendor had supported blits for 64-bit formats so far,
which is why this problem had gone unnoticed. The assert happens when
the test tries to generate data for the source image, using the color
gradient fill strategy.

The solution is to generate 32-bit values and store them as 64-bit
values. This is not ideal because coverage is not going to be full.
However, it's a good first approach to at least remove the assertions.
In CTS:

- Code to properly store 64-bit values as floats is tricky and will
  clearly result in precision issues, specially if blitting.
- convertSatRte does not support 64-bit values and such code would be
  a bit tricky to write.
- Even if we fixed convertSatRte,
  CopiesAndBlittingTestInstance::generateBuffer currently truncates
  channel bit width to 32-bit to avoid precision issues.
- PixelBufferAccess has no methods to store 64-bit ints in images, which
  would allow us to approach part of the problem from a different side.

For the reasons above, fix the framework functions even if the high 64
bits of each image pixel will be left as zeros.

Affects:
dEQP-VK.api.copy_and_blit.core.blit_image.*

Components: Framework
VK-GL-CTS issue: 5784

Change-Id: I0439487e5d5c254ae2d00d456fd8e7c9d5bedc47