Add compile-time option to log test case source code origin
This should help locate the source code for a particular test case. No
test results should be affected by this change, which is also opt-in (so
by default no test case source is logged).

To enable the new functionality, configure the project with the
following option:

-DDEQP_LOG_NODE_SOURCE=1

After that, TestResults.qpa should contain a CaseSource attribute for
each test case, like this:

<TestCaseResult CasePath="dEQP-VK.api.smoke.triangle"
  CaseSource="/SOME/PATH/external/vulkancts/modules/vulkan/api/vktApiSmokeTests.cpp:862"
  Version="0.3.4" CaseType="SelfValidate">

The functionality relies on C++23's <stacktrace>, and some additional
changes are needed to remove warnings or fix compilation errors when
enabling C++23 mode on the compiler. A recent compiler and C++ standard
library are needed.

There is a memory usage and CPU usage penalty paid when the source code
file and line are stored in each test node.

For example, ./deqp-vk --deqp-duplicate-case-name-check=disable
--deqp-runmode=stdout-caselist takes longer to run with the option
enabled: 40 seconds vs 1 minute on my system.

Components: Framework, Vulkan, OpenGL
VK-GL-CTS issue: 5796

Change-Id: Ia4a996e098f6c7a8ddd97d44dfb506c0916e35d1