Add testing of VK_ARM_tensors extension
Initial batch of tensor tests.

The SimpleAllocator device memory allocation class has been modified to
always map the allocation if the allocation is host visible. We do this
because we want to simplify using host access when possible and fall
back on staging buffers when the allocation is not host visible.

New Tests:
dEQP-VK.tensor.*

Components: Vulkan

VK-GL-CTS issue: 6032

Change-Id: Id40fc4be4663576636220cc00d0675c8f47ba6fb