Exclude fetching and building the VVL by default
A new --include-vvl option is added to fetch_sources.py. When used, the
layers will be downloaded and the post-checkout procedure will be run,
downloading and preparing their internal dependencies. In addition, the
layers will be built by default unless a specific target is selected
that excludes them.

However, if --include-vvl is not used, the layers will not be downloaded
nor built by default.

VK-GL-CTS issue: 5561

Change-Id: I0a469ae13291ac05d097f72a96e45adabd858f42