OpConstantNull and cooperative matrices
The calculations are performed in two stages for each
configuration available in the system. First, a normal
matrix multiplication and addition is performed; second,
the specified matrix is nulled and the multiplication
and addition is performed again. Each stage can fail,
so the results can be viewed in the logs.

VK-GL-CTS Issue: 5625

Components: Vulkan

New tests:
dEQP-VK.compute.pipeline.cooperative_matrix.op_constant_null.*

Change-Id: I3a168708a5b2707e8659e5c51bab0bed0e0af1f2