VideoEncodeTestsAV1: keep same name for input filename
As the input filename can be a downloaded content with
fetch_optional_video_encode_samples.py, it has to keep
a generic name.
If the content does not exist, the content will be generated
and the file will be kept to avoid concurrency issues.

Affects:
dEQP-VK.video.encode.av1.*

Components: Vulkan

VK-GL-CTS issue: 5539

Change-Id: I30a550ba6116fa05379abce23e39caa6a24cc763