Respect strictness when using PRIMITIVESTRICTNESS_IGNORE
It's not up to the test to decide if lines are being rasterized as
strict lines or not. This depends on the device-limits. In the case
where the strictness has been passed as STRICT or NONSTRICT, it should
already have been verified elsewhere that the strictness matches.

However, in the case where the strictness is set as IGNORE, we still
need to respect the setting in the device-limits; we can't just treat
these as non-strict like we currently do.

So let's respect the actual strictness here, instead of treating these
as non-strict.

Without this, we incorrectly expect axis-aligned endpoints on the
stipples when testing stippled lines on implementations with strictLines,
instead of line-aligned like we should.

Affected tests:
dEQP-VK.rasterization.*.line*

Components: Vulkan

Change-Id: I22a6180b3138d579bee0dea66d91690b0863128a