Temporarily disable sign-comparison warnings for vulkan-video-samples
vulkan-video-samples refer to the main branch of the Vulkan headers (and loader), instead of using a fixed commit hash.
This means that as the external Vulkan headers are updated (out of vulkan-video-samples control), compilation issues may be introduced.

In this case luckly it's just sign-comparison warnings that were tread as errors. This change ignores them, but a better solution needs to be found.

VK-GL-CTS issue: 5881

Change-Id: I1e1d2d9e37458968c9a614e373ee5bbee20f2df4