Test DGC IES with null pSetLayoutsInfo
When creating an indirect execution set with shader objects, we need to
pass the set layouts for each shader stage as part of the creation info.
This commits add a test that makes the info null, which would force the
implementation to get that information from the initial set of
VkShaderEXT objects the set is created with.

New test:
dEQP-VK.dgc.ext.compute.misc.null_set_layouts_info

Components: Vulkan
VK-GL-CTS issue: 5877

Change-Id: I3e4fd906c165e544e05c3ac0b882f862ca0e23ae