Fix uniform buffer count for descriptor set alloc test
The device_object_reservation.limits.request_count.descriptorset
test allocates 5 descriptor sets, each of which contain a single uniform
buffer descriptor. The descriptor pool is only allocated with space for
one uniform buffer, so this causes issues when allocating.

To fix this, reserve space in the descriptor pool that accounts for all
of the descriptor sets that the test allocates.

This bug went undetected previously on nvidia's implementation because
enough extra space (for padding error) was added to the descriptor pool
so that the allocations passed, but recent driver changes have caused
that extra space to shrink enough that these allocations fail.

VK-GL-CTS issue: 5904
Components: VulkanSC
Affects: VKSC.sc.device_object_reservation.limits.request_count.descriptorset

Change-Id: I2db3ec6d6a87526fca9c10c798641d2a7cc28bb7