Fix memory leak in descriptor indexing tests
PixelBufferAccessAllocation was trying to allocate an array of bytes,
and then use an std::vector<unsigned char> to track that allocation, but
it was doing so incorrectly. The code was creating a new (actually
unused) vector and leaking the original allocation instead.

This commit stores the newly allocated array to an
std::unique_ptr<unsigned char[]> that will be destroyed with the object.

Affects:
dEQP-VK.descriptor_indexing.*

VK-GL-CTS issue: 5918
VK-GL-CTS public issue: 531

Components: Vulkan
Change-Id: If179ef0d913f578fc1271aeafdbf3c17ef97377d